+++ lib/fortnox/structs/invoice_row.rb
@@ -4,10 +4,10 @@ module Fortnox
   module Structs
     class InvoiceRow < DocumentRow
       # PriceExcludingVAT Price per unit excluding VAT.
-      attribute? :price_excluding_vat, Types::Coercible::Float.optional, :read_only
+      attribute? :price_excluding_vat, Types::UnsizedFloat, :read_only
 
       # TotalExcludingVAT Total amount for the row excluding VAT.
-      attribute? :total_excluding_vat, Types::Coercible::Float.optional, :read_only
+      attribute? :total_excluding_vat, Types::UnsizedFloat, :read_only
     end
   end
 end